requirements.txt
requests
pydub
openai
assemblyai
elevenlabs
python-dotenv
